Hw04 auth

+++ model/contacts/removeContact.js
@@ -1,8 +1,9 @@
 const {contact} = require('./schema/contactSchema')
 
-const removeContact = async (contactId) => {
-  await contact.findByIdAndDelete(contactId)
-  return {message: "contact is deleted"}
+const removeContact = async (contactId, owner) => {
+  const removedContact = await contact.findByIdAndDelete({ _id: contactId, owner })
+   
+  return removedContact
 }
 
 module.exports = {
